+++ src/delay.rs
@@ -3,7 +3,7 @@
 //! ## Usage
 //!
 //! If the actual sleep duration is not important, simply create a
-//! [`MockNoop`](struct.MockNoop.html) instance. There will be no actual
+//! [`NoopDelay`](struct.NoopDelay.html) instance. There will be no actual
 //! delay. This is useful for fast tests, where you don't actually need to wait
 //! for the hardware.
 //!
@@ -18,16 +18,16 @@ use std::time::Duration;
 use embedded_hal::blocking::delay;
 
 /// A `Delay` implementation that does not actually block.
-pub struct MockNoop;
+pub struct NoopDelay;
 
-impl MockNoop {
-    /// Create a new `MockNoop` instance.
+impl NoopDelay {
+    /// Create a new `NoopDelay` instance.
     pub fn new() -> Self {
-        MockNoop
+        NoopDelay
     }
 }
 
-impl Default for MockNoop {
+impl Default for NoopDelay {
     fn default() -> Self {
         Self::new()
     }
@@ -35,7 +35,7 @@ impl Default for MockNoop {
 
 macro_rules! impl_noop_delay_us {
     ($type:ty) => {
-        impl delay::DelayUs<$type> for MockNoop {
+        impl delay::DelayUs<$type> for NoopDelay {
             /// A no-op delay implementation.
             fn delay_us(&mut self, _n: $type) {}
         }
@@ -49,7 +49,7 @@ impl_noop_delay_us!(u64);
 
 macro_rules! impl_noop_delay_ms {
     ($type:ty) => {
-        impl delay::DelayMs<$type> for MockNoop {
+        impl delay::DelayMs<$type> for NoopDelay {
             /// A no-op delay implementation.
             fn delay_ms(&mut self, _n: $type) {}
         }