+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - The minimal supported Rust version (MSRV) is specified in the `Cargo.toml` to
   offer clearer error messages to consumers with outdated Rust versions
+- Renamed `delay::MockNoop` to `delay::NoopDelay`.
 
 
 ## 0.9.0 - 2023-01-07